User login, wishlist post and UI correction

+++ app/src/main/java/com/makaan/response/wishlist/WishList.java
@@ -7,7 +7,7 @@ import com.makaan.response.project.Project;
  * Created by sunil on 25/01/16.
  */
 public class WishList {
-    public Integer listingId, projectId, wishListId;
+    public Long listingId, projectId, wishListId;
     public String cityLabel, unitName, builderName, projectName;
     public Project project;
     public ListingDetail listing;
